Rewrite son to work based on generators.
This is neccessary to be able to parse files that are
large. Using generators means that we don't need to first
load the full file into memory, we do everything on the fly.

The outward-facing interface of son.load and son.dump
remains unchanged, but we add son.read, which returns
an iterator over entries in the file.

In order to do this cleanly I rewrote almost everything,
except the tests. ;) I think the reorganised code is
mostly self-documenting, but happy to discuss any
particular changes.